// src/constants/interfaces.ts
export type Page = {
  title: string;
  link?: string;
  selected?: boolean;
  onPress?: () => void;
  hidden?: boolean;
  children?: Page[];
};
// ...
export type StateSetter<T> = React.Dispatch<React.SetStateAction<T>>;